Sorting: Reorganised book sort code to its own directory
